added the empty checks to update page

+++ app/controllers/admin/tips_controller.rb
@@ -57,14 +57,26 @@ class Admin::TipsController < ApplicationController
   def update
     @tip = Tip.find params[:id]
     if @tip
-      @tip.update_attributes! params[:tip]
-      
-      # remove all categories with this tip's id, in lieu of adding the selected ones
-      CategoryTip.destroy_all :tip_id => params[:id]
-      if params[:categories]
-        params[:categories].each do |name|
-          # checked is always "1" in here, i.e. only checked categories are passed
-          CategoryTip.create! :tip_id => params[:id], :category_id => Category.find_by_name(name).id
+      # Check for empty title or content fields
+      if @tip.title == ""
+        flash[:error] = "Please specify a title for the tip."
+        # TODO: return to the "new" page and keep the form fields intact
+        redirect_to new_admin_tip_path
+        return
+      elsif @tip.content == ""
+        flash[:error] = "Please specify some content for this tip."
+        # TODO: return to the "new" page and keep the form fields intact
+        redirect_to new_admin_tip_path
+        return
+      else
+        @tip.update_attributes! params[:tip]
+        # remove all categories with this tip's id, in lieu of adding the selected ones
+        CategoryTip.destroy_all :tip_id => params[:id]
+        if params[:categories]
+          params[:categories].each do |name|
+            # checked is always "1" in here, i.e. only checked categories are passed
+            CategoryTip.create! :tip_id => params[:id], :category_id => Category.find_by_name(name).id
+          end
         end
       end
       flash[:notice] = "#{@tip.title} was succuessfully updated."
